HAI-1710 Sanitize special chars in attachment filenames
Sanitize any special characters (`\/:*?"<>|`) in attachment file names.
The special characters are replaced with underscores (`_`).

Any `"` in the filename is encoded to `%22` by the browser. Treat `%22`
in the filename like it was `"`, i.e. change the substring to a single
underscore.

Change id field in attachment metadata classes to non-nullable. There's
no reason to create a metadata presentation of an attachment if it's not
saved to database.